+++ src/comps/Product/Tooltip/ProductHistory.svelte
@@ -1,25 +1,27 @@
 <script lang="ts">
   import { storedStates } from '@stores';
-  import { jsVoid, pad, getStateAgo } from '@utils';
+  import { jsVoid, pad, getStateAgo, stopClick } from '@utils';
+  import ProductHistoryEntry from './ProductHistoryEntry.svelte';
 
   export let product: any;
 
+  let nextVisible = false;
   let fullVisible = false;
   let fullHistory: any;
   let splittedHistory: any;
   let states: any;
   const historyChunks = 3;
+  const moreChunks = 20;
 
   storedStates.subscribe(store => (states = store));
 
   const getFormattedDate = timestamp => {
     const date = new Date(timestamp);
-    const year = date.getFullYear();
     const month = pad(date.getMonth() + 1);
     const day = pad(date.getDate());
     const hours = pad(date.getHours());
     const minutes = pad(date.getMinutes());
-    return `${day}.${month}.${year} ${hours}:${minutes}`;
+    return `${day}.${month} ${hours}:${minutes}`;
   };
 
   const getStateLabel = stateId => {
@@ -27,44 +29,63 @@
   };
 
   fullHistory = Object.entries(product.history).reverse();
+  let latestYear = 0;
   splittedHistory = {
     first: fullHistory.map(([timestamp, stateId], i) => {
       timestamp = parseInt(timestamp);
       const compareDate = fullHistory[i - 1] ? parseInt(fullHistory[i - 1][0]) : 0;
+      const entryYear = new Date(timestamp).getFullYear();
+      let newYear = false;
+      if (entryYear !== latestYear) {
+        if (latestYear !== 0) {
+          newYear = true;
+        }
+        latestYear = entryYear;
+      }
       return {
         formattedDate: getFormattedDate(timestamp),
         stateId,
         label: getStateLabel(stateId),
         ago: getStateAgo(product, stateId, timestamp, compareDate, i),
+        ...(newYear && { newYear: entryYear }),
       };
     }),
   };
 
   if (fullHistory.length > historyChunks) {
-    splittedHistory.last = splittedHistory.first.slice(historyChunks, fullHistory.length);
+    splittedHistory.last = splittedHistory.first.slice(moreChunks, fullHistory.length);
+    splittedHistory.next = splittedHistory.first.slice(historyChunks, moreChunks);
     splittedHistory.first = splittedHistory.first.slice(0, historyChunks);
   }
 </script>
 
 <div class="history">
   {#each splittedHistory.first as entry}
-    <div>
-      <span class="date">
-        {entry.formattedDate}
-      </span>
-      - <span>{entry.label}</span>
-      {#if entry.ago}&nbsp;
-        <b>
-          {entry.ago}
-        </b>
-      {/if}
-    </div>
+    <ProductHistoryEntry {entry} />
   {/each}
-  {#if splittedHistory.last}
+  {#if splittedHistory.next}
+    {#if !nextVisible}
+      <a
+        on:click={e => {
+          stopClick(e);
+          nextVisible = true;
+        }}
+        href={jsVoid}
+        class="link"
+      >
+        ... Mehr anzeigen
+      </a>
+    {:else}
+      {#each splittedHistory.next as entry}
+        <ProductHistoryEntry {entry} />
+      {/each}
+    {/if}
+  {/if}
+  {#if splittedHistory.last.length > 0 && nextVisible}
     {#if !fullVisible}
       <a
         on:click={e => {
-          e.stopPropagation();
+          stopClick(e);
           fullVisible = true;
         }}
         href={jsVoid}
@@ -74,7 +95,7 @@
       </a>
     {:else}
       {#each splittedHistory.last as entry}
-        {entry.formattedDate} - {entry.label}{#if entry.ago}&nbsp;<b>{entry.ago}</b>{/if}<br />
+        <ProductHistoryEntry {entry} />
       {/each}
     {/if}
   {/if}
@@ -83,13 +104,5 @@
 <style lang="scss">
   .history {
     user-select: none;
-
-    .date {
-      display: block;
-
-      @media (min-width: 400px) {
-        display: inline-block;
-      }
-    }
   }
 </style>